Allow for running ros nodes on different routes
