+++ gui/ros_gui/gui/configpanel/configpanel.py
@@ -4,13 +4,11 @@ import os
 import time
 import json
 
-import rclpy
 from rcl_interfaces.msg import ParameterType, Parameter
 from rcl_interfaces.srv import SetParametersAtomically, GetParameters, ListParameters
 from rclpy.node import Node
-from rclpy.executors import ExternalShutdownException
 
-from nicegui import Client, ui, app, ui_run
+from nicegui import Client, ui, app
 from .local_file_picker import local_file_picker
 
 CONF_PATH = os.path.join(os.path.expanduser('~'), ".astazero/ATOS/conf")
@@ -353,26 +351,4 @@ class ConfigPanelNode(Node):
         self.parameters = {}
         threading.Thread(target=self.get_parameters_list, args=(self.client_list,)).start()
         ui.open('/config')
-
-def ros_main() -> None:
-    """ Initializes the ROS node and spins it.
-    """
-    rclpy.init()
-    node = ConfigPanelNode()
-    try:
-        rclpy.spin(node)
-    except ExternalShutdownException:
-        pass
-
-def main() -> None:
-    """ Main function for ROS.
-        NOTE: This function is defined as the ROS entry point in setup.py, but it's empty to enable NiceGUI auto-reloading
-    """
-    pass
-
-#Starting the ros node in a thread managed by nicegui. It will restarted with "on_startup" after a reload.
-#It has to be in a thread, since NiceGUI wants the main thread for itself.
-app.on_startup(lambda: threading.Thread(target=ros_main).start())
-
-ui_run.APP_IMPORT_STRING = f'{__name__}:app'  # ROS2 uses a non-standard module name, so we need to specify it here
-ui.run(uvicorn_reload_dirs=str(Path(__file__).parent.resolve()), favicon='🤖', port=3001, title='Config Panel')
+
