+++ gui/ros_gui/gui/controlpanel/controlpanel.py
@@ -1,13 +1,10 @@
-import threading
-from pathlib import Path
 from atos_interfaces.srv import *
 
 import rclpy
 from std_msgs.msg import Empty
-from rclpy.executors import ExternalShutdownException
 from rclpy.node import Node
 
-from nicegui import Client, app, ui, ui_run, run
+from nicegui import Client, ui
 
 QOS = rclpy.qos.QoSProfile(depth=10)
 
@@ -23,9 +20,8 @@ OBC_STATES = {
         8: 'ABORTING'
     }
 class ControlPanelNode(Node):
-
     def __init__(self) -> None:
-        super().__init__('nicegui')
+        super().__init__('control_panel')
         self.initPub = self.create_publisher(Empty, '/atos/init', QOS)
         self.connectPub = self.create_publisher(Empty, '/atos/connect', QOS)
         self.disconnectPub = self.create_publisher(Empty, '/atos/disconnect', QOS)
@@ -45,6 +41,11 @@ class ControlPanelNode(Node):
         self.lost_connection = True
 
         with Client.auto_index_client:
+            pass
+
+
+        @ui.page(path='/control')
+        def render_page():
             with ui.row().bind_visibility_from(self, 'lost_connection'):
                 ui.label('Lost connection to ATOS...')
             with ui.row():
@@ -63,19 +64,16 @@ class ControlPanelNode(Node):
                 ui.button('Reset Test Objects', on_click=lambda: self.resetTestObjectsPub.publish(Empty()), color='grey')
                 ui.button('Reload Object Settings', on_click=lambda: self.reloadObjectSettingsPub.publish(Empty()), color='grey')
 
+
     def get_object_control_state_callback(self):
         # Call the service
-        service_timeout_counter = 0
         while not self.get_object_control_state_client.wait_for_service(timeout_sec=1.0):
-            if service_timeout_counter > 5:
-                self.get_logger().info('object control state service not available, giving up')
-                return
-            service_timeout_counter += 1
-            self.get_logger().info('object control state service not available, waiting again...')
+            self.get_logger().info('service not available, waiting again...')
             self.lost_connection = True
         self.lost_connection = False
         future = self.get_object_control_state_client.call_async(self.OBC_state_req)
         future.add_done_callback(lambda future: self.get_object_control_state_callback_done(future))
+
 
     def get_object_control_state_callback_done(self, future):
         try:
@@ -84,25 +82,3 @@ class ControlPanelNode(Node):
             self.get_logger().info('Service call failed %r' % (e,))
         else:
             self.OBC_state["state"] = OBC_STATES[response.state]
-    
-
-
-def main() -> None:
-    # NOTE: This function is defined as the ROS entry point in setup.py, but it's empty to enable NiceGUI auto-reloading
-    pass
-
-
-def ros_main() -> None:
-    rclpy.init()
-    node = ControlPanelNode()
-    try:
-        rclpy.spin(node)
-    except ExternalShutdownException:
-        pass
-
-#Starting the ros node in a thread managed by nicegui. It will restarted with "on_startup" after a reload.
-#It has to be in a thread, since NiceGUI wants the main thread for itself.
-app.on_startup(lambda: threading.Thread(target=ros_main).start())
-
-ui_run.APP_IMPORT_STRING = f'{__name__}:app'  # ROS2 uses a non-standard module name, so we need to specify it here
-ui.run(uvicorn_reload_dirs=str(Path(__file__).parent.resolve()), favicon='🤖', port=3000)
